+++ tests/testthat/test-shoji.R
@@ -44,7 +44,18 @@ test_that("ShojiCatalog", {
 with(fake.HTTP, {
     full.urls <- DatasetCatalog(crGET("/api/datasets.json"))
     rel.urls <- DatasetCatalog(crGET("/api/datasets-relative-urls.json"))
-    expect_identical(urls(full.urls), urls(rel.urls))
+    test_that("urls() method returns absolute URLs", {
+        expect_identical(urls(full.urls), urls(rel.urls))
+    })
+
+    test_that("shojiURL", {
+        ds <- loadDataset("test ds")
+        expect_identical(shojiURL(ds, "catalogs", "variables"),
+            "/api/datasets/dataset1/variables.json")
+        expect_error(shojiURL(ds, "catalogs", "NOTACATALOG"),
+            paste0("No URL ", dQuote("NOTACATALOG"), " in collection ",
+            dQuote("catalogs")))
+    })
 })
 
 if (run.integration.tests) {